Redesign PayloadBuffer and add MM and stream interfaces to it
I moved PayloadBuffer to a separate project, because it will be packet into
a Qsys module. I also redesigned it, so that the code looks more readable.
I removed byteCount field from it, because it seems more optimal to keep it
in the packet header. I added two ST and one MM interface to it, so that it
can be used with SG-DMA. However, now I think that DMA is too complicated
and consumes a lot of resources, so I am planning to implement my own small
DMA inside the payload buffer.